Rearranged code for readability
placed all of the science and housekeeping functions together along with the rx callback to reduce the amount of file jumping needed when following the code. Also renamed files to reflect the code rearrangement changes
