+++ src/sci_hk.h
@@ -3,7 +3,6 @@
 
 #include  <stdint.h>
 
-#include "can_queue.h"
 #include "obc.h"
 
 
@@ -12,8 +11,13 @@
 #define EPS_HK_BLOCK_SIZE 100
 #define PAY_HK_BLOCK_SIZE 2
 
+#define SCI_BLOCK_SIZE 218
+#define SCI_TYPE 0 // kamran will define in his code, delete this then
+
 // CAN callbacks
 void req_hk_timer_callback();
+void data_rx_mob_callback(const uint8_t* data, uint8_t len);
+void receive_science(uint8_t field_num, uint8_t* data);
 void receive_hk(uint8_t board_num, uint8_t field_num, uint8_t* payload);
 void write_to_flash(uint8_t board_num, uint8_t field_num, uint8_t* data);
