+++ src/timer_callbacks.h
@@ -1,0 +1,11 @@
+#ifndef SCIENCE_H
+#define SCIENCE_H
+
+#include  <stdint.h>
+
+#include "obc.h"
+
+void req_sci_timer_callback();
+void req_hk_timer_callback();
+
+#endif
